test: move data to common location
